refactor(ui): resolve half-controlled status prop and apply SRP on notification utils
- Resolve half-controlled state architectural contradiction by making the component purely uncontrolled on its initial status/notifications props, as per React state best practices. This fully eliminates any risk of infinite render loops and dual-render paint flashes.
- Apply Single Responsibility Principle (SRP) by extracting NotificationItem type, defaultMockNotifications data, and getNotificationContent rendering mapping logic to src/components/layout/Header/notificationUtils.tsx, decoupling domain business logic from view rendering.
